bug fix: step can be greater than 127

+++ app/inc/data.h
@@ -54,7 +54,7 @@ typedef struct CONTROL_T {
     uint8_t properties_mask;
     uint8_t properties;
     float value, minimum, maximum;
-    int8_t step, steps;
+    int32_t step, steps;
     uint8_t controls_count, control_index;
     uint8_t scale_points_count;
     scale_point_t **scale_points;
